+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Changelog
 
 ### 4.0.2 (Next)
+* [#248](https://github.com/alexa-js/alexa-app/pull/248): Add support for intent and slot confirmationStatus - [@sbonami]((https://github.com/sbonami).
 * Your contribution here
 
 ### 4.0.1 (May 17, 2017)