+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Changelog
 
-### 4.0.2 (Next)
+### 4.1.0 (Next)
 * [#97](https://github.com/alexa-js/alexa-app/pull/97): Added support for custom directives - [@ajcrites](https://github.com/ajcrites).
 * [#248](https://github.com/alexa-js/alexa-app/pull/248): Add support for intent and slot `confirmationStatus` - [@sbonami]((https://github.com/sbonami).
 * [#250](https://github.com/alexa-js/alexa-app/pull/250): Add support for handling and/or delegating Dialog - [@sbonami]((https://github.com/sbonami).